Remove the use of the springRepulsionFactor
This was used to push nodes away from the center of the edge between
other nodes. The goal of this was to reduce edge crossing, but it just
pushes edge crossings to the the side, instead of the middle of edges
(not removing the edge at all). This is usually less readable because
edge crossings are closer to actual nodes, and it also makes
calculations longer.

Signed-off-by: NathanDissoubray <nathan.dissoubray@rte-france.com>