Remove spring middle repulsion (#694)
* Remove the use of the springRepulsionFactor; this was used to push nodes away from the center of the edge between
other nodes. The goal of this was to reduce edge crossing, but it just
pushes edge crossings to the the side, instead of the middle of edges
(not removing the edge at all). This is usually less readable because
edge crossings are closer to actual nodes, and it also makes
calculations longer.
* Modify impacted code and tests
* Clean some test resources

Signed-off-by: NathanDissoubray <nathan.dissoubray@rte-france.com>